+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-* add possibility to convert hex to in in `calculator` processor with new added function `from_hex`
+* add possibility to convert hex to int in `calculator` processor with new added function `from_hex`
 
 ### Improvements
 ### Bugfix